FocusZen: Fixup bunch of stuff

+++ backend/src/main/java/com/focuszen/dto/BreakDTO.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class BreakDTO {
     private Long id;
     private String type;
-    private int startTime;
+    private long startTime;
     private int durationInMinutes;
     private LocalDateTime timestamp;
 }
