+++ backend/src/main/java/com/focuszen/dto/MindfulnessDTO.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class MindfulnessDTO {
     private Long id;
     private String activityType;
-    private int startTime;
-    private int durationInMinutes;
+    private long startTime;
+    private long durationInMinutes;
     private LocalDateTime timestamp;
 }
